third_party/blink/web_tests/html/sections/numbered-header-element-expected.txt

Test that any numbered header element end tag can close any other open numbered header element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<h1> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h1>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h1> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h2>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h1> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h3>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h1> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h4>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h1> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h5>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h1> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h2> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h3> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h4> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h5> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
<h6> closes <h6>:
PASS test2.parentNode.id == "test1" is false
PASS test2.parentNode.id == "test0" is true
PASS successfullyParsed is true

TEST COMPLETE
